requirements.txt
setuptools>=65
whisper
#torch
#torchvision
#torchaudio
#torch torchvision torchaudio --index-url https://download.pytorch.org/whl/cu118
torch --index-url https://download.pytorch.org/whl/cpu
torchaudio>=2
faster-whisper==1.1.1
transformers
nltk
git+https://github.com/m-bain/whisperx.git
google-generativeai
openai
python-multipart
streamlit
google-generativeai
openai
langchain_community
pypdfium2